Specifies a bootstrap card

Specifies a bootstrap card

cards(
  data,
  title = NULL,
  text = NULL,
  image = NULL,
  link = NULL,
  footer = NULL,
  header = NULL,
  tags = NULL,
  layout = "label-below",
  width = 3,
  spacing = 2,
  breakpoint = 4,
  label_colour = "#ffffffaa",
  border_colour = "#808080",
  border_width = 1,
  rounding = 3
)

Arguments

data
Data

title
Title for the card

text
Text for the card

image
Path to the card image

link
URL to link to from title and image

footer
Card footer

header
Card header

tags
Tags to be assigned to each card

layout
Card layout is a string, defaults to "label-below" (see details)

width
Card width is an integer between 1 and 5

spacing
Spacing between cards is an integer between 0 and 5

breakpoint
Number between 1 and 5, controlling label size on horizontal and inset layouts

label_colour
Colour applied to the card label

border_colour
Colour applied to the card border

border_width
Width of card border is an integer between 0 and 5

rounding
Amount of rounding on card corners is an integer between 0 and 5

Value

A "shiny.tag" object